+++ cli/src/definitions.ts
@@ -96,12 +96,16 @@ export interface IOSConfig extends PlatformConfig {
   readonly cordovaPluginsDirAbs: string;
   readonly minVersion: string;
   readonly podPath: string;
-  readonly webDir: string;
-  readonly webDirAbs: string;
+  readonly webDir: Promise<string>;
+  readonly webDirAbs: Promise<string>;
   readonly nativeProjectDir: string;
   readonly nativeProjectDirAbs: string;
   readonly nativeTargetDir: string;
   readonly nativeTargetDirAbs: string;
+  readonly nativeXcodeProjDir: string;
+  readonly nativeXcodeProjDirAbs: string;
+  readonly nativeXcodeWorkspaceDir: Promise<string>;
+  readonly nativeXcodeWorkspaceDirAbs: Promise<string>;
 }
 
 export type WebConfig = PlatformConfig;