+++ cli/src/tasks/copy.ts
@@ -66,7 +66,7 @@ export async function copy(
     }
 
     if (platformName === config.ios.name) {
-      await copyWebDir(config, config.ios.webDirAbs);
+      await copyWebDir(config, await config.ios.webDirAbs);
       await copyCapacitorConfig(config, config.ios.nativeTargetDirAbs);
       const cordovaPlugins = await getCordovaPlugins(config, platformName);
       await handleCordovaPluginsJS(cordovaPlugins, config, platformName);